ViewModelFactoryの削除、Hilt-Moduleの分割、postValueからsetValueへ変更
