+++ app/src/main/java/com/example/twitchapp/di/NetworkModule.kt
@@ -1,0 +1,30 @@
+package com.example.twitchapp.di
+
+import com.example.twitchapp.model.api.TwitchApi
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+import javax.inject.Singleton
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NetworkModule {
+    private val BASE_URL = "https://api.twitch.tv/kraken/"
+    private val moshi = Moshi.Builder().add(KotlinJsonAdapterFactory()).build()
+
+    @Singleton
+    @Provides
+    fun provideRetrofitInstance(): TwitchApi {
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .build()
+            .create(TwitchApi::class.java)
+    }
+}
